+++ app/src/main/java/srimani7/apps/feedfly/navigation/FavoriteScreen.kt
@@ -48,7 +48,7 @@ import coil.compose.AsyncImage
 import coil.request.ImageRequest
 import srimani7.apps.feedfly.R
 import srimani7.apps.feedfly.database.FavoriteArticle
-import srimani7.apps.feedfly.rss.DateParser
+import srimani7.apps.rssparser.DateParser
 import srimani7.apps.feedfly.ui.ArticleFavoriteToggle
 import srimani7.apps.feedfly.ui.DescriptionText
 import srimani7.apps.feedfly.ui.ShowImageDialog
